Creación de UserManager Funcional

+++ Maquina 1/src/main/java/SantanderSAS/Controller/UserManager.java
@@ -3,7 +3,7 @@ package SantanderSAS.Controller;
 import SantanderSAS.Model.Domain.User;
 import SantanderSAS.Model.Repository.UserRepository;
 
-public class    UserManager {
+public class UserManager {
     private UserRepository userRepository;
 
     public UserManager(UserRepository userRepository){
@@ -14,19 +14,27 @@ public class    UserManager {
         userRepository.addUser(user);
     }
 
-    public void RemoveUser(User user){
+    // En UserManager.java
+    public void RemoveUser(String id){
+    User user = userRepository.getUserById(id);
+    if (user != null) {
         userRepository.removeUser(user);
     }
+}
 
     public void EditUser(User user){
+    User existingUser = userRepository.getUserById(user.getID());
+    if (existingUser != null) {
         userRepository.editUser(user);
     }
+}
 
     public User getUser(String username){
         return userRepository.getUser(username);
     }
 
     public User[] GetUsers(){
-        return null;
+        User[] users = userRepository.getUsers();
+        return users != null ? users : new User[0];
     }
 }
